Everything Heroku! 1. Clean up. 2. Rearrange folders. 3. Add README images for highlights and Presentation deck. 4. Extra styling including favicon and About page. 5. Notebook updates with testing codes (commented out). 6. Delete experimenting files inclidung app_danni for presentation purpose. 7. Update app.py to use pickle and joblib to load vectorizer, scaler and classifier in order to optimize predicting speed. 8. All set up docs for Heroku Deployment: requirements.txt, runtime.txt, Profile, .slugignore, nltk.txt, etc. 9. Alternative app.py that fits the model directly in flask is saved as 'app_fitted_alternative.py' 10. Any other minor changes...
