+++ flask_app/templates/about.html
@@ -1,0 +1,151 @@
+<!DOCTYPE html>
+<html lang="en-us">
+
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0, maximum-scale=1.0, user-scalable=no" />
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>About</title>
+    <!--Bootswatch CSS Theme-->
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css') }}">
+    <!-- <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"> -->
+    <!-- Our CSS  -->
+    <link rel="stylesheet" href="{{ url_for('static', filename='css/about.css') }}">
+    <!-- Favicon -->
+    <link rel="icon" href="{{ url_for('static', filename='img/movie_rating_predictor.ico') }}">
+  </head>
+
+  <body>
+    <!--Navigation Bar-->
+    <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
+      <div class="container-fluid">
+        <a class="navbar-brand" href="/">Movie Rating Predictor</a>
+        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
+          <span class="navbar-toggler-icon"></span>
+        </button>
+        <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
+          <div class="navbar-nav">
+            <a class="nav-link" href="/about">About</a>
+            <!-- <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true"></a> -->
+          </div>
+        </div>
+      </div>
+    </nav>
+
+
+    
+    <!-- Container -->
+    <div class="container">
+
+      <div class="row justify-content-center top-row">
+        <div class="col">
+          <h2>About the Team</h2>
+          <p>Click <a href="https://github.com/DPintacasi/finalproject" class="text-secondary link-underline" target="_blank">here</a> to visit our Github repo.</p>
+          <br>
+        </div>
+      </div>
+
+      <div class="row">
+        <div class="col-md-4">
+          <div class="card border-dark mb-3">
+            <h3 class="card-header header-small">Alannah<br>Fleet</h3>
+            <div>
+                <img src="{{ url_for('static', filename='img/alannah_fleet.png') }}" alt="Alannah Fleet" width="100%">
+            </div>
+            <div class="card-body">
+              <p class="card-text">
+                Alannah is.... 
+                intro... 
+                intro...
+              </p>
+            </div>
+            <ul class="list-group list-group-flush">
+              <li class="list-group-item">Favorite movie: <br><em>movie_name</em></li>
+              <li class="list-group-item">Hobbies: <br>some hobbies</li>
+            </ul>
+            <div class="card-body">
+                <span class="badge rounded-pill bg-dark">
+                    <a href="https://github.com/alannahmarie" class="link-underline" target="_blank">Github</a>
+                </span>
+                <span class="badge rounded-pill bg-dark">
+                    <a href="https://www.linkedin.com/in/alannah-fleet-106129101/" class="link-underline" target="_blank">LinkedIn</a>
+                </span>
+            </div>
+          </div>
+        </div>
+        <div class="col-md-4">
+          <div class="card border-dark mb-3">
+            <h3 class="card-header header-small">Danielle<br>Pintacasi</h3>
+            <div>
+              <img src="{{ url_for('static', filename='img/danni_pintacasi.png') }}" alt="Danni Pintacasi" width="100%">
+            </div>
+            <div class="card-body">
+              <p class="card-text">
+                Danni is.... 
+                intro... 
+                intro...
+              </p>
+            </div>
+            <ul class="list-group list-group-flush">
+              <li class="list-group-item">Favorite movie: <br><em>movie_name</em></li>
+              <li class="list-group-item">Hobbies: <br>some hobbies</li>
+            </ul>
+            <div class="card-body">
+              <span class="badge rounded-pill bg-dark">
+                <a href="https://github.com/DPintacasi" class="link-underline" target="_blank">Github</a>
+              </span>
+              <span class="badge rounded-pill bg-dark">
+                <a href="https://www.linkedin.com/in/danielle-pintacasi-49640a1a9/" class="link-underline" target="_blank">LinkedIn</a>
+              </span>
+            </div>
+          </div>
+        </div>
+        <div class="col-md-4">
+          <div class="card border-dark mb-3">
+            <h3 class="card-header header-small">Tianchi<br>Zhang</h3>
+            <div>
+              <img src="{{ url_for('static', filename='img/tianchi_zhang.png') }}" alt="Tianchi Zhang" width="100%">
+            </div>
+            <div class="card-body">
+              <p class="card-text">
+                Tianchi is a multidisciplinary analyst who's passionate about empowering decision 
+                makings with actionable data and creating exceptional products. He currently works in 
+                the generic pharma industry as a product analyst.
+              </p>
+            </div>
+            <ul class="list-group list-group-flush">
+              <li class="list-group-item">Favorite movie: <br><em>2001: A Space Odyssey</em></li>
+              <li class="list-group-item">Hobbies: <br>Basketball, film, space</li>
+            </ul>
+            <div class="card-body">
+              <span class="badge rounded-pill bg-dark">
+                <a href="https://github.com/ztcnrh" class="link-underline" target="_blank">Github</a>
+              </span>
+              <span class="badge rounded-pill bg-dark">
+                <a href="https://www.linkedin.com/in/tianchi-zhang-tc/" class="link-underline" target="_blank">LinkedIn</a>
+              </span>
+            </div>
+          </div>
+        </div>
+      </div>
+    </div>
+
+    
+
+
+
+
+
+    <!-- D3 CDN -->
+    <script src="https://d3js.org/d3.v5.min.js"></script>
+    <!-- JS -->
+    <!-- <script type="text/javascript" src="{{ url_for('static', filename='js/about.js') }}"></script> -->
+    
+
+    <script src="https://cdn.jsdelivr.net/npm/@popperjs/core@2.9.1/dist/umd/popper.min.js" integrity="sha384-SR1sx49pcuLnqZUnnPwx6FCym0wLsk5JZuNx2bPPENzswTNFaQU1RDvt3wT4gWFG" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.0.0-beta3/dist/js/bootstrap.min.js" integrity="sha384-j0CNLUeiqtyaRmlzUHCPZ+Gy5fQu0dQ6eZ/xAww941Ai1SxSY+0EQqNXNE6DZiVc" crossorigin="anonymous"></script>
+  
+  
+  </body>
+
+</html>
