+++ flask_app/templates/index.html
@@ -9,9 +9,10 @@
     <!--Bootswatch CSS Theme-->
     <link rel="stylesheet" href="{{ url_for('static', filename='css/bootstrap.min.css') }}">
     <!-- <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous"> -->
-    
     <!-- Our CSS  -->
     <link rel="stylesheet" href="{{ url_for('static', filename='css/index.css') }}">
+    <!-- Favicon -->
+    <link rel="icon" href="{{ url_for('static', filename='img/movie_rating_predictor.ico') }}">
 
     <style>
       /* Media query rating output */
@@ -44,12 +45,12 @@
     <nav class="navbar navbar-expand-lg navbar-dark bg-primary">
       <div class="container-fluid">
         <a class="navbar-brand" href="/">Movie Rating Predictor</a>
-        <!-- <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
+        <button class="navbar-toggler" type="button" data-bs-toggle="collapse" data-bs-target="#navbarNavAltMarkup" aria-controls="navbarNavAltMarkup" aria-expanded="false" aria-label="Toggle navigation">
           <span class="navbar-toggler-icon"></span>
-        </button> -->
+        </button>
         <div class="collapse navbar-collapse" id="navbarNavAltMarkup">
           <div class="navbar-nav">
-            <!-- <a class="nav-link" aria-current="page" href="/">Home</a> -->
+            <a class="nav-link" href="/about">About</a>
             <!-- <a class="nav-link disabled" href="#" tabindex="-1" aria-disabled="true"></a> -->
           </div>
         </div>
