--- flask_app/templates/rating.html
@@ -1,49 +1,0 @@
-{% extends "index.html" %}
-{% block content %}
-   <!-- Output Area -->
-   <div class="col-4">
-    {% if predictions == 0%}
-    <div class="card text-white bg-success mb-3" style="max-width: 20rem;">
-      <div class="card-header">Prediction</div>
-      <div class="card-body">
-        <h4 class="card-title">Your movie is rated:</h4>
-        <h1>G</h1>
-      </div>
-    </div>
-    {% elif predictions == 1%}
-    <div class="card bg-light mb-3" style="max-width: 20rem;">
-      <div class="card-header">Prediction</div>
-      <div class="card-body">
-        <h4 class="card-title">Your movie is rated:</h4>
-        <h1>PG</h1>
-      </div>
-    </div>
-    {% elif predictions == 2%}
-    <div class="card text-white bg-warning mb-3" style="max-width: 20rem;">
-      <div class="card-header">Prediction</div>
-      <div class="card-body">
-        <h4 class="card-title">Your movie is rated:</h4>
-        <h1>PG-13</h1>
-      </div>
-    </div>
-    {% elif predictions == 3%}
-    <div class="card text-white bg-danger mb-3" style="max-width: 20rem;">
-      <div class="card-header">Prediction</div>
-      <div class="card-body">
-        <h4 class="card-title">Your movie is rated:</h4>
-        <h1>R</h1>         
-      </div>
-    </div>   
-    {% endif %}
-
-    <!-- Placeholder card -->
-    <!-- <div class="card text-white bg-secondary mb-3" style="max-width: 20rem;">
-    <div class="card-header">Prediction</div>
-    <div class="card-body">
-      <h4 class="card-title">Your movie is:</h4>
-      <h1>NOT YET RATED!</h1>
-    </div>
-  </div> -->
-
-  {% endblock %}
-</div>
