+++ nltk.txt
@@ -1,0 +1,2 @@
+stopwords
+wordnet
