+++ requirements.txt
@@ -1,0 +1,33 @@
+certifi==2021.5.30
+click==8.0.1
+Flask==2.0.1
+gunicorn==20.1.0
+importlib-metadata==4.5.0
+itsdangerous==2.0.1
+Jinja2==3.0.1
+# joblib @ file:///tmp/build/80754af9/joblib_1613502643832/work
+joblib==1.0.1
+MarkupSafe==2.0.1
+nltk==3.6.2
+# numexpr @ file:///opt/concourse/worker/volumes/live/1df078ad-9528-4b49-6728-a99d819a0f1f/volume/numexpr_1618856547208/work
+numexpr==2.7.3
+# numpy @ file:///opt/concourse/worker/volumes/live/fe7fac38-ab01-46f0-4a7d-7569ed91a56f/volume/numpy_and_numpy_base_1620831185168/work
+numpy==1.20.2
+pandas==1.2.4
+# python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
+python-dateutil==2.8.1
+# pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
+pytz==2021.1
+regex==2021.4.4
+# scikit-learn @ file:///opt/concourse/worker/volumes/live/4b416fcf-2a1b-4a30-781a-c588a35f6ca3/volume/scikit-learn_1614446663856/work
+scikit-learn==0.24.1
+# scipy @ file:///opt/concourse/worker/volumes/live/28879f84-8cc6-46f1-5728-21456d2e76ef/volume/scipy_1618855943934/work
+scipy==1.6.2
+# six @ file:///tmp/build/80754af9/six_1623709665295/work
+six==1.16.0
+# threadpoolctl @ file:///tmp/tmp9twdgx9k/threadpoolctl-2.1.0-py3-none-any.whl
+threadpoolctl==2.1.0
+tqdm==4.61.1
+typing-extensions==3.10.0.0
+Werkzeug==2.0.1
+zipp==3.4.1
