+++ run.sh
@@ -1,0 +1,1 @@
+FLASK_APP=flask_app/app.py flask run
